chore(core): move the shared vitest setup into tools/vitest
At the workspace root, vite resolved the setup against the root solution tsconfig and walked its references, reading all 114 project tsconfigs on every run. A leaf tsconfig beside the files cuts that to one, so the workspace:test inputs no longer have to name them.
